Refactor Milkdrop renderer adapter helpers
