+++ assets/js/milkdrop/renderer-adapter.ts
@@ -1,24 +1,22 @@
-import type { Camera, Scene } from 'three';
+import type {
+  Camera,
+  Points,
+  PointsMaterial,
+  Scene,
+  ShaderMaterial,
+} from 'three';
 import {
   AdditiveBlending,
   BufferGeometry,
-  Color,
-  DoubleSide,
   DynamicDrawUsage,
   Float32BufferAttribute,
   Group,
-  Line,
+  type Line,
   LineBasicMaterial,
-  LineLoop,
   LineSegments,
   Mesh,
   MeshBasicMaterial,
-  NormalBlending,
-  Path,
   PlaneGeometry,
-  Points,
-  PointsMaterial,
-  ShaderMaterial,
   Shape,
   ShapeGeometry,
   Sphere,
@@ -34,7 +32,14 @@ import {
   updateBorderLine as updateBorderLineHelper,
 } from './renderer-helpers/border-renderer';
 import { buildFeedbackCompositeState as buildFeedbackCompositeStateHelper } from './renderer-helpers/feedback-composite';
+import { renderMesh as renderMeshHelper } from './renderer-helpers/mesh-renderer';
 import { renderMotionVectors as renderMotionVectorsHelper } from './renderer-helpers/motion-vector-renderer';
+import {
+  syncInterpolatedProceduralCustomWaveObject,
+  syncInterpolatedProceduralWaveObject,
+  syncProceduralCustomWaveObject,
+  syncProceduralWaveObject,
+} from './renderer-helpers/procedural-wave-renderer';
 import {
   createShapeObject as createShapeObjectHelper,
   renderShapeGroup as renderShapeGroupHelper,
@@ -55,13 +60,9 @@ import type {
   MilkdropFeedbackCompositeState,
   MilkdropFeedbackManager,
   MilkdropFeedbackSetRenderTarget,
-  MilkdropGpuFieldExpression,
-  MilkdropGpuFieldProgramDescriptor,
-  MilkdropGpuFieldSignalInputs,
   MilkdropGpuGeometryHints,
   MilkdropGpuInteractionTransform,
   MilkdropProceduralCustomWaveVisual,
-  MilkdropProceduralFieldTransformVisual,
   MilkdropProceduralWaveVisual,
   MilkdropRendererAdapter,
   MilkdropRenderPayload,
@@ -197,1124 +198,10 @@ export const WEBGPU_MILKDROP_BACKEND_BEHAVIOR: MilkdropBackendBehavior = {
 };
 
 const SHARED_GEOMETRY_FLAG = 'milkdropSharedGeometry';
-const PROCEDURAL_MESH_BOUNDS_RADIUS = Math.SQRT2 * 2;
-const PROCEDURAL_MOTION_VECTOR_BOUNDS_RADIUS = Math.SQRT2 * 2.35;
-const PROCEDURAL_WAVE_BOUNDS_RADIUS = Math.SQRT2 * 2.2;
 const BACKGROUND_GEOMETRY = markSharedGeometry(new PlaneGeometry(6.4, 6.4));
 const polygonFillGeometryCache = new Map<number, ShapeGeometry>();
 const polygonOutlineGeometryCache = new Map<string, BufferGeometry>();
-const proceduralMeshGeometryCache = new Map<number, BufferGeometry>();
-const proceduralMotionVectorGeometryCache = new Map<string, BufferGeometry>();
-const proceduralWaveGeometryCache = new Map<number, BufferGeometry>();
 
-type ProceduralFieldUniformState = {
-  zoom: { value: number };
-  zoomExponent: { value: number };
-  rotation: { value: number };
-  warp: { value: number };
-  warpAnimSpeed: { value: number };
-  centerX: { value: number };
-  centerY: { value: number };
-  scaleX: { value: number };
-  scaleY: { value: number };
-  translateX: { value: number };
-  translateY: { value: number };
-  time: { value: number };
-  trebleAtt: { value: number };
-  tint: { value: Color };
-  alpha: { value: number };
-  signalTime: { value: number };
-  signalFrame: { value: number };
-  signalFps: { value: number };
-  signalBass: { value: number };
-  signalMid: { value: number };
-  signalMids: { value: number };
-  signalTreble: { value: number };
-  signalBassAtt: { value: number };
-  signalMidAtt: { value: number };
-  signalMidsAtt: { value: number };
-  signalTrebleAtt: { value: number };
-  signalBeat: { value: number };
-  signalBeatPulse: { value: number };
-  signalRms: { value: number };
-  signalVol: { value: number };
-  signalMusic: { value: number };
-  signalWeightedEnergy: { value: number };
-};
-
-type ProceduralFieldVisualWithSignals =
-  MilkdropProceduralFieldTransformVisual & {
-    signals: MilkdropGpuFieldSignalInputs;
-  };
-
-type ProceduralInteractionUniformState = {
-  interactionOffsetX: { value: number };
-  interactionOffsetY: { value: number };
-  interactionRotation: { value: number };
-  interactionScale: { value: number };
-  interactionAlpha: { value: number };
-};
-
-function createProceduralFieldUniformState() {
-  return {
-    zoom: { value: 1 },
-    zoomExponent: { value: 1 },
-    rotation: { value: 0 },
-    warp: { value: 0 },
-    warpAnimSpeed: { value: 1 },
-    centerX: { value: 0 },
-    centerY: { value: 0 },
-    scaleX: { value: 1 },
-    scaleY: { value: 1 },
-    translateX: { value: 0 },
-    translateY: { value: 0 },
-    time: { value: 0 },
-    trebleAtt: { value: 0 },
-    tint: { value: new Color(1, 1, 1) },
-    alpha: { value: 1 },
-    signalTime: { value: 0 },
-    signalFrame: { value: 0 },
-    signalFps: { value: 60 },
-    signalBass: { value: 0 },
-    signalMid: { value: 0 },
-    signalMids: { value: 0 },
-    signalTreble: { value: 0 },
-    signalBassAtt: { value: 0 },
-    signalMidAtt: { value: 0 },
-    signalMidsAtt: { value: 0 },
-    signalTrebleAtt: { value: 0 },
-    signalBeat: { value: 0 },
-    signalBeatPulse: { value: 0 },
-    signalRms: { value: 0 },
-    signalVol: { value: 0 },
-    signalMusic: { value: 0 },
-    signalWeightedEnergy: { value: 0 },
-  } satisfies ProceduralFieldUniformState;
-}
-
-function createProceduralInteractionUniformState() {
-  return {
-    interactionOffsetX: { value: 0 },
-    interactionOffsetY: { value: 0 },
-    interactionRotation: { value: 0 },
-    interactionScale: { value: 1 },
-    interactionAlpha: { value: 1 },
-  } satisfies ProceduralInteractionUniformState;
-}
-
-const PROCEDURAL_INTERACTION_SHADER_CHUNK = `
-  vec2 applyMilkdropInteraction(vec2 point) {
-    vec2 scaled = point * interactionScale;
-    float cosRot = cos(interactionRotation);
-    float sinRot = sin(interactionRotation);
-    return vec2(
-      scaled.x * cosRot - scaled.y * sinRot + interactionOffsetX,
-      scaled.x * sinRot + scaled.y * cosRot + interactionOffsetY
-    );
-  }
-`;
-
-const PROCEDURAL_FIELD_PROGRAM_SIGNAL_PARAMETERS = `
-  float signalTimeValue,
-  float signalFrameValue,
-  float signalFpsValue,
-  float signalBassValue,
-  float signalMidValue,
-  float signalMidsValue,
-  float signalTrebleValue,
-  float signalBassAttValue,
-  float signalMidAttValue,
-  float signalMidsAttValue,
-  float signalTrebleAttValue,
-  float signalBeatValue,
-  float signalBeatPulseValue,
-  float signalRmsValue,
-  float signalVolValue,
-  float signalMusicValue,
-  float signalWeightedEnergyValue
-`;
-
-const PROCEDURAL_FIELD_PROGRAM_SHADER_HELPERS = `
-  float milkdropBool(float value) {
-    return abs(value) > 0.000001 ? 1.0 : 0.0;
-  }
-
-  float milkdropSelect(bool condition, float whenTrue, float whenFalse) {
-    return condition ? whenTrue : whenFalse;
-  }
-
-  float milkdropBitOr(float left, float right) {
-    return float(int(left) | int(right));
-  }
-
-  float milkdropBitAnd(float left, float right) {
-    return float(int(left) & int(right));
-  }
-
-  float milkdropBitNot(float value) {
-    return float(~int(value));
-  }
-
-  float milkdropFrac(float value) {
-    return value - floor(value);
-  }
-
-  float milkdropSigmoid(float value, float slope) {
-    return 1.0 / (1.0 + exp(-value * slope));
-  }
-
-  float milkdropIf(float condition, float whenTrue, float whenFalse) {
-    return abs(condition) > 0.000001 ? whenTrue : whenFalse;
-  }
-
-  float milkdropAbove(float left, float right) {
-    return left > right ? 1.0 : 0.0;
-  }
-
-  float milkdropBelow(float left, float right) {
-    return left < right ? 1.0 : 0.0;
-  }
-
-  float milkdropEqual(float left, float right) {
-    return abs(left - right) <= 0.000001 ? 1.0 : 0.0;
-  }
-
-  float milkdropNormalizeTransformCenterX(float value) {
-    return value >= 0.0 && value <= 1.0 ? (value - 0.5) * 2.0 : value;
-  }
-
-  float milkdropNormalizeTransformCenterY(float value) {
-    return value >= 0.0 && value <= 1.0 ? (0.5 - value) * 2.0 : value;
-  }
-
-  float milkdropDenormalizeTransformCenterX(float value) {
-    return value * 0.5 + 0.5;
-  }
-
-  float milkdropDenormalizeTransformCenterY(float value) {
-    return 0.5 - value * 0.5;
-  }
-`;
-
-function gpuFieldVarName(name: string) {
-  switch (name) {
-    case 'zoom':
-      return 'fieldZoom';
-    case 'zoomexp':
-      return 'fieldZoomExponent';
-    case 'rot':
-      return 'fieldRotation';
-    case 'warp':
-      return 'fieldWarp';
-    case 'cx':
-      return 'fieldCenterX';
-    case 'cy':
-      return 'fieldCenterY';
-    case 'sx':
-      return 'fieldScaleX';
-    case 'sy':
-      return 'fieldScaleY';
-    case 'dx':
-      return 'fieldTranslateX';
-    case 'dy':
-      return 'fieldTranslateY';
-    default:
-      return `field_${name}`;
-  }
-}
-
-function gpuFieldIdentifierToShaderSource(name: string) {
-  switch (name) {
-    case 'pi':
-      return '3.141592653589793';
-    case 'e':
-      return '2.718281828459045';
-    case 'time':
-      return 'signalTimeValue';
-    case 'frame':
-      return 'signalFrameValue';
-    case 'fps':
-      return 'signalFpsValue';
-    case 'bass':
-      return 'signalBassValue';
-    case 'mid':
-      return 'signalMidValue';
-    case 'mids':
-      return 'signalMidsValue';
-    case 'treble':
-      return 'signalTrebleValue';
-    case 'bassAtt':
-      return 'signalBassAttValue';
-    case 'midAtt':
-      return 'signalMidAttValue';
-    case 'midsAtt':
-      return 'signalMidsAttValue';
-    case 'trebleAtt':
-      return 'signalTrebleAttValue';
-    case 'beat':
-      return 'signalBeatValue';
-    case 'beatPulse':
-      return 'signalBeatPulseValue';
-    case 'rms':
-      return 'signalRmsValue';
-    case 'vol':
-      return 'signalVolValue';
-    case 'music':
-      return 'signalMusicValue';
-    case 'weightedEnergy':
-      return 'signalWeightedEnergyValue';
-    default:
-      return gpuFieldVarName(name);
-  }
-}
-
-function buildGpuFieldExpressionShaderSource(
-  expression: MilkdropGpuFieldExpression,
-): string {
-  switch (expression.type) {
-    case 'literal':
-      return Number.isFinite(expression.value)
-        ? expression.value.toString()
-        : '0.0';
-    case 'identifier':
-      return gpuFieldIdentifierToShaderSource(expression.name);
-    case 'unary': {
-      const operand = buildGpuFieldExpressionShaderSource(expression.operand);
-      if (expression.operator === '!') {
-        return `(milkdropBool(${operand}) > 0.5 ? 0.0 : 1.0)`;
-      }
-      return `(${expression.operator}${operand})`;
-    }
-    case 'binary': {
-      const left = buildGpuFieldExpressionShaderSource(expression.left);
-      const right = buildGpuFieldExpressionShaderSource(expression.right);
-      switch (expression.operator) {
-        case '^':
-          return `pow(${left}, ${right})`;
-        case '%':
-          return `(abs(${right}) <= 0.000001 ? 0.0 : mod(${left}, ${right}))`;
-        case '|':
-          return `milkdropBitOr(${left}, ${right})`;
-        case '&':
-          return `milkdropBitAnd(${left}, ${right})`;
-        case '<':
-        case '<=':
-        case '>':
-        case '>=':
-        case '==':
-        case '!=':
-          return `(${left} ${expression.operator} ${right} ? 1.0 : 0.0)`;
-        case '&&':
-          return `((milkdropBool(${left}) > 0.5 && milkdropBool(${right}) > 0.5) ? 1.0 : 0.0)`;
-        case '||':
-          return `((milkdropBool(${left}) > 0.5 || milkdropBool(${right}) > 0.5) ? 1.0 : 0.0)`;
-        default:
-          return `(${left} ${expression.operator} ${right})`;
-      }
-    }
-    case 'call': {
-      const args = expression.args.map(buildGpuFieldExpressionShaderSource);
-      switch (expression.name) {
-        case 'mod':
-        case 'fmod':
-          return `(abs(${args[1]}) <= 0.000001 ? 0.0 : mod(${args[0]}, ${args[1]}))`;
-        case 'mix':
-        case 'lerp':
-          return `mix(${args[0]}, ${args[1]}, ${args[2]})`;
-        case 'int':
-          return `floor(${args[0]})`;
-        case 'sqr':
-          return `((${args[0]}) * (${args[0]}))`;
-        case 'sigmoid':
-          return `milkdropSigmoid(${args[0]}, ${args[1]})`;
-        case 'sign':
-          return `sign(${args[0]})`;
-        case 'bor':
-          return `milkdropBitOr(${args[0]}, ${args[1]})`;
-        case 'band':
-          return `milkdropBitAnd(${args[0]}, ${args[1]})`;
-        case 'bnot':
-          return `milkdropBitNot(${args[0]})`;
-        case 'atan2':
-          return `atan(${args[0]}, ${args[1]})`;
-        case 'frac':
-          return `milkdropFrac(${args[0]})`;
-        case 'if':
-          return `milkdropIf(${args[0]}, ${args[1]}, ${args[2]})`;
-        case 'above':
-          return `milkdropAbove(${args[0]}, ${args[1]})`;
-        case 'below':
-          return `milkdropBelow(${args[0]}, ${args[1]})`;
-        case 'equal':
-          return `milkdropEqual(${args[0]}, ${args[1]})`;
-        default:
-          return `${expression.name}(${args.join(', ')})`;
-      }
-    }
-  }
-}
-
-function buildProceduralFieldProgramShaderChunk(
-  program: MilkdropGpuFieldProgramDescriptor | null | undefined,
-) {
-  if (!program) {
-    return `
-      ${PROCEDURAL_FIELD_PROGRAM_SHADER_HELPERS}
-
-      vec2 milkdropTransformPointWithParams(
-        vec2 source,
-        float paramZoom,
-        float paramZoomExponent,
-        float paramRotation,
-        float paramWarp,
-        float paramWarpAnimSpeed,
-        float paramCenterX,
-        float paramCenterY,
-        float paramScaleX,
-        float paramScaleY,
-        float paramTranslateX,
-        float paramTranslateY,
-        ${PROCEDURAL_FIELD_PROGRAM_SIGNAL_PARAMETERS}
-      ) {
-        float radius = length(source);
-        float angle = atan(source.y, source.x) + paramRotation;
-        float transformedX =
-          (source.x - paramCenterX) * paramScaleX + paramCenterX + paramTranslateX;
-        float transformedY =
-          (source.y - paramCenterY) * paramScaleY + paramCenterY + paramTranslateY;
-        float ripple = sin(
-          radius * 12.0 +
-          signalTimeValue * (0.6 + signalTrebleAttValue) * (0.35 + paramWarpAnimSpeed)
-        ) * paramWarp * 0.08;
-        float radiusNormalized = clamp(radius / 1.41421356237, 0.0, 1.0);
-        float zoomScale = pow(
-          max(paramZoom, 0.0001),
-          pow(max(paramZoomExponent, 0.0001), radiusNormalized * 2.0 - 1.0)
-        );
-        vec2 warped = vec2(
-          (transformedX + cos(angle * 3.0) * ripple) * zoomScale,
-          (transformedY + sin(angle * 4.0) * ripple) * zoomScale
-        );
-        float cosRot = cos(paramRotation);
-        float sinRot = sin(paramRotation);
-        return vec2(
-          warped.x * cosRot - warped.y * sinRot,
-          warped.x * sinRot + warped.y * cosRot
-        );
-      }
-    `;
-  }
-
-  const temporaryDeclarations = program.temporaries
-    .map((temporary) => `float ${gpuFieldVarName(temporary)} = 0.0;`)
-    .join('\n        ');
-  const statementCode = program.statements
-    .map(
-      (statement) =>
-        `${gpuFieldVarName(statement.target)} = ${buildGpuFieldExpressionShaderSource(
-          statement.expression,
-        )};`,
-    )
-    .join('\n        ');
-  return `
-    ${PROCEDURAL_FIELD_PROGRAM_SHADER_HELPERS}
-
-    vec2 milkdropTransformPointWithParams(
-      vec2 source,
-      float paramZoom,
-      float paramZoomExponent,
-      float paramRotation,
-      float paramWarp,
-      float paramWarpAnimSpeed,
-      float paramCenterX,
-      float paramCenterY,
-      float paramScaleX,
-      float paramScaleY,
-      float paramTranslateX,
-      float paramTranslateY,
-      ${PROCEDURAL_FIELD_PROGRAM_SIGNAL_PARAMETERS}
-    ) {
-      float field_x = source.x;
-      float field_y = source.y;
-      float field_rad = length(source);
-      float field_ang = atan(source.y, source.x);
-      float fieldZoom = paramZoom;
-      float fieldZoomExponent = paramZoomExponent;
-      float fieldRotation = paramRotation;
-      float fieldWarp = paramWarp;
-      float fieldCenterX = milkdropDenormalizeTransformCenterX(paramCenterX);
-      float fieldCenterY = milkdropDenormalizeTransformCenterY(paramCenterY);
-      float fieldScaleX = paramScaleX;
-      float fieldScaleY = paramScaleY;
-      float fieldTranslateX = paramTranslateX * 0.5;
-      float fieldTranslateY = paramTranslateY * 0.5;
-      ${temporaryDeclarations}
-      ${statementCode}
-
-      float normalizedCenterX = milkdropNormalizeTransformCenterX(fieldCenterX);
-      float normalizedCenterY = milkdropNormalizeTransformCenterY(fieldCenterY);
-      float angle = field_ang + fieldRotation;
-      float translatedX =
-        (field_x - normalizedCenterX) * fieldScaleX +
-        normalizedCenterX +
-        fieldTranslateX * 2.0;
-      float translatedY =
-        (field_y - normalizedCenterY) * fieldScaleY +
-        normalizedCenterY +
-        fieldTranslateY * 2.0;
-      float ripple = sin(
-        field_rad * 12.0 +
-        signalTimeValue * (0.6 + signalTrebleAttValue) * (0.35 + paramWarpAnimSpeed)
-      ) * fieldWarp * 0.08;
-      float radiusNormalized = clamp(field_rad / 1.41421356237, 0.0, 1.0);
-      float zoomScale = pow(
-        max(fieldZoom, 0.0001),
-        pow(max(fieldZoomExponent, 0.0001), radiusNormalized * 2.0 - 1.0)
-      );
-      float px = (translatedX + cos(angle * 3.0) * ripple) * zoomScale;
-      float py = (translatedY + sin(angle * 4.0) * ripple) * zoomScale;
-      float cosRot = cos(fieldRotation);
-      float sinRot = sin(fieldRotation);
-      return vec2(
-        px * cosRot - py * sinRot,
-        px * sinRot + py * cosRot
-      );
-    }
-  `;
-}
-
-function createProceduralMeshMaterial(
-  program?: MilkdropGpuFieldProgramDescriptor | null,
-) {
-  const uniforms = {
-    ...createProceduralFieldUniformState(),
-    ...createProceduralInteractionUniformState(),
-  };
-  const fieldProgramShader = buildProceduralFieldProgramShaderChunk(program);
-  return new ShaderMaterial({
-    uniforms,
-    userData: {
-      fieldProgramSignature: program?.signature ?? 'default',
-    },
-    transparent: true,
-    vertexShader: `
-      attribute vec3 sourcePosition;
-      uniform float zoom;
-      uniform float zoomExponent;
-      uniform float rotation;
-      uniform float warp;
-      uniform float warpAnimSpeed;
-      uniform float time;
-      uniform float trebleAtt;
-      uniform float interactionOffsetX;
-      uniform float interactionOffsetY;
-      uniform float interactionRotation;
-      uniform float interactionScale;
-      uniform float signalTime;
-      uniform float signalFrame;
-      uniform float signalFps;
-      uniform float signalBass;
-      uniform float signalMid;
-      uniform float signalMids;
-      uniform float signalTreble;
-      uniform float signalBassAtt;
-      uniform float signalMidAtt;
-      uniform float signalMidsAtt;
-      uniform float signalTrebleAtt;
-      uniform float signalBeat;
-      uniform float signalBeatPulse;
-      uniform float signalRms;
-      uniform float signalVol;
-      uniform float signalMusic;
-      uniform float signalWeightedEnergy;
-      ${fieldProgramShader}
-      ${PROCEDURAL_INTERACTION_SHADER_CHUNK}
-
-      void main() {
-        vec2 transformed = applyMilkdropInteraction(
-          milkdropTransformPointWithParams(
-            sourcePosition.xy,
-            zoom,
-            zoomExponent,
-            rotation,
-            warp,
-            warpAnimSpeed,
-            centerX,
-            centerY,
-            scaleX,
-            scaleY,
-            translateX,
-            translateY,
-            signalTime,
-            signalFrame,
-            signalFps,
-            signalBass,
-            signalMid,
-            signalMids,
-            signalTreble,
-            signalBassAtt,
-            signalMidAtt,
-            signalMidsAtt,
-            signalTrebleAtt,
-            signalBeat,
-            signalBeatPulse,
-            signalRms,
-            signalVol,
-            signalMusic,
-            signalWeightedEnergy
-          )
-        );
-        gl_Position = projectionMatrix * modelViewMatrix * vec4(
-          transformed.xy,
-          sourcePosition.z,
-          1.0
-        );
-      }
-    `,
-    fragmentShader: `
-      uniform vec3 tint;
-      uniform float alpha;
-      uniform float interactionAlpha;
-
-      void main() {
-        gl_FragColor = vec4(tint, alpha * interactionAlpha);
-      }
-    `,
-  });
-}
-
-function createProceduralMotionVectorMaterial(
-  program?: MilkdropGpuFieldProgramDescriptor | null,
-) {
-  const uniforms = {
-    ...createProceduralFieldUniformState(),
-    ...createProceduralInteractionUniformState(),
-    sourceOffsetX: { value: 0 },
-    sourceOffsetY: { value: 0 },
-    explicitLength: { value: 0 },
-    legacyControls: { value: 0 },
-    previousZoom: { value: 1 },
-    previousZoomExponent: { value: 1 },
-    previousRotation: { value: 0 },
-    previousWarp: { value: 0 },
-    previousWarpAnimSpeed: { value: 1 },
-    previousCenterX: { value: 0 },
-    previousCenterY: { value: 0 },
-    previousScaleX: { value: 1 },
-    previousScaleY: { value: 1 },
-    previousTranslateX: { value: 0 },
-    previousTranslateY: { value: 0 },
-    previousSourceOffsetX: { value: 0 },
-    previousSourceOffsetY: { value: 0 },
-    previousExplicitLength: { value: 0 },
-    blendMix: { value: 1 },
-    previousSignalTime: { value: 0 },
-    previousSignalFrame: { value: 0 },
-    previousSignalFps: { value: 60 },
-    previousSignalBass: { value: 0 },
-    previousSignalMid: { value: 0 },
-    previousSignalMids: { value: 0 },
-    previousSignalTreble: { value: 0 },
-    previousSignalBassAtt: { value: 0 },
-    previousSignalMidAtt: { value: 0 },
-    previousSignalMidsAtt: { value: 0 },
-    previousSignalTrebleAtt: { value: 0 },
-    previousSignalBeat: { value: 0 },
-    previousSignalBeatPulse: { value: 0 },
-    previousSignalRms: { value: 0 },
-    previousSignalVol: { value: 0 },
-    previousSignalMusic: { value: 0 },
-    previousSignalWeightedEnergy: { value: 0 },
-  };
-  const fieldProgramShader = buildProceduralFieldProgramShaderChunk(program);
-  return new ShaderMaterial({
-    uniforms,
-    userData: {
-      fieldProgramSignature: program?.signature ?? 'default',
-    },
-    transparent: true,
-    vertexShader: `
-      attribute vec3 sourcePosition;
-      attribute float endpointWeight;
-      uniform float zoom;
-      uniform float zoomExponent;
-      uniform float rotation;
-      uniform float warp;
-      uniform float warpAnimSpeed;
-      uniform float centerX;
-      uniform float centerY;
-      uniform float scaleX;
-      uniform float scaleY;
-      uniform float translateX;
-      uniform float translateY;
-      uniform float time;
-      uniform float trebleAtt;
-      uniform float sourceOffsetX;
-      uniform float sourceOffsetY;
-      uniform float explicitLength;
-      uniform float interactionOffsetX;
-      uniform float interactionOffsetY;
-      uniform float interactionRotation;
-      uniform float interactionScale;
-      uniform float interactionAlpha;
-      uniform float previousZoom;
-      uniform float previousZoomExponent;
-      uniform float previousRotation;
-      uniform float previousWarp;
-      uniform float previousWarpAnimSpeed;
-      uniform float previousCenterX;
-      uniform float previousCenterY;
-      uniform float previousScaleX;
-      uniform float previousScaleY;
-      uniform float previousTranslateX;
-      uniform float previousTranslateY;
-      uniform float previousSourceOffsetX;
-      uniform float previousSourceOffsetY;
-      uniform float previousExplicitLength;
-      uniform float blendMix;
-      varying float vAlpha;
-      uniform float signalTime;
-      uniform float signalFrame;
-      uniform float signalFps;
-      uniform float signalBass;
-      uniform float signalMid;
-      uniform float signalMids;
-      uniform float signalTreble;
-      uniform float signalBassAtt;
-      uniform float signalMidAtt;
-      uniform float signalMidsAtt;
-      uniform float signalTrebleAtt;
-      uniform float signalBeat;
-      uniform float signalBeatPulse;
-      uniform float signalRms;
-      uniform float signalVol;
-      uniform float signalMusic;
-      uniform float signalWeightedEnergy;
-      uniform float previousSignalTime;
-      uniform float previousSignalFrame;
-      uniform float previousSignalFps;
-      uniform float previousSignalBass;
-      uniform float previousSignalMid;
-      uniform float previousSignalMids;
-      uniform float previousSignalTreble;
-      uniform float previousSignalBassAtt;
-      uniform float previousSignalMidAtt;
-      uniform float previousSignalMidsAtt;
-      uniform float previousSignalTrebleAtt;
-      uniform float previousSignalBeat;
-      uniform float previousSignalBeatPulse;
-      uniform float previousSignalRms;
-      uniform float previousSignalVol;
-      uniform float previousSignalMusic;
-      uniform float previousSignalWeightedEnergy;
-      ${fieldProgramShader}
-      ${PROCEDURAL_INTERACTION_SHADER_CHUNK}
-
-      void main() {
-        vec2 currentSource = clamp(
-          sourcePosition.xy + vec2(sourceOffsetX, sourceOffsetY),
-          vec2(-1.0),
-          vec2(1.0)
-        );
-        vec2 previousSource = clamp(
-          sourcePosition.xy + vec2(previousSourceOffsetX, previousSourceOffsetY),
-          vec2(-1.0),
-          vec2(1.0)
-        );
-        vec2 current = milkdropTransformPointWithParams(
-          currentSource,
-          zoom,
-          zoomExponent,
-          rotation,
-          warp,
-          warpAnimSpeed,
-          centerX,
-          centerY,
-          scaleX,
-          scaleY,
-          translateX,
-          translateY,
-          signalTime,
-          signalFrame,
-          signalFps,
-          signalBass,
-          signalMid,
-          signalMids,
-          signalTreble,
-          signalBassAtt,
-          signalMidAtt,
-          signalMidsAtt,
-          signalTrebleAtt,
-          signalBeat,
-          signalBeatPulse,
-          signalRms,
-          signalVol,
-          signalMusic,
-          signalWeightedEnergy
-        );
-        vec2 previous = milkdropTransformPointWithParams(
-          previousSource,
-          previousZoom,
-          previousZoomExponent,
-          previousRotation,
-          previousWarp,
-          previousWarpAnimSpeed,
-          previousCenterX,
-          previousCenterY,
-          previousScaleX,
-          previousScaleY,
-          previousTranslateX,
-          previousTranslateY,
-          previousSignalTime,
-          previousSignalFrame,
-          previousSignalFps,
-          previousSignalBass,
-          previousSignalMid,
-          previousSignalMids,
-          previousSignalTreble,
-          previousSignalBassAtt,
-          previousSignalMidAtt,
-          previousSignalMidsAtt,
-          previousSignalTrebleAtt,
-          previousSignalBeat,
-          previousSignalBeatPulse,
-          previousSignalRms,
-          previousSignalVol,
-          previousSignalMusic,
-          previousSignalWeightedEnergy
-        );
-        vec2 blendedCurrent = mix(previous, current, blendMix);
-        vec2 blendedSource = mix(previousSource, currentSource, blendMix);
-        vec2 delta = clamp(
-          (blendedCurrent - blendedSource) * 1.35,
-          vec2(-0.24),
-          vec2(0.24)
-        );
-        float explicitLengthMixed = mix(
-          previousExplicitLength,
-          explicitLength,
-          blendMix
-        );
-        float magnitude = length(delta);
-        if (explicitLengthMixed > 0.0001 && magnitude > 0.0001) {
-          delta = delta / magnitude * explicitLengthMixed;
-          magnitude = length(delta);
-        }
-        vec2 renderPoint = mix(
-          blendedCurrent - delta * 0.45,
-          blendedCurrent + delta,
-          endpointWeight
-        );
-        renderPoint = applyMilkdropInteraction(renderPoint);
-        vAlpha =
-          alpha *
-          interactionAlpha *
-          clamp(0.75 + magnitude * 2.2, 0.0, 1.0) *
-          step(0.002, magnitude);
-        gl_Position = projectionMatrix * modelViewMatrix * vec4(
-          renderPoint.xy,
-          sourcePosition.z,
-          1.0
-        );
-      }
-    `,
-    fragmentShader: `
-      uniform vec3 tint;
-      varying float vAlpha;
-
-      void main() {
-        if (vAlpha <= 0.0) {
-          discard;
-        }
-        gl_FragColor = vec4(tint, vAlpha);
-      }
-    `,
-  });
-}
-
-function createProceduralWaveMaterial() {
-  return new ShaderMaterial({
-    uniforms: {
-      mode: { value: 0 },
-      centerX: { value: 0 },
-      centerY: { value: 0 },
-      scale: { value: 0.34 },
-      mystery: { value: 0 },
-      signalTime: { value: 0 },
-      beatPulse: { value: 0 },
-      trebleAtt: { value: 0 },
-      tint: { value: new Color(1, 1, 1) },
-      alpha: { value: 1 },
-      previousCenterX: { value: 0 },
-      previousCenterY: { value: 0 },
-      previousScale: { value: 0.34 },
-      previousMystery: { value: 0 },
-      previousSignalTime: { value: 0 },
-      previousBeatPulse: { value: 0 },
-      previousTrebleAtt: { value: 0 },
-      blendMix: { value: 1 },
-      ...createProceduralInteractionUniformState(),
-    },
-    transparent: true,
-    vertexShader: `
-      attribute float sampleT;
-      attribute float sampleValue;
-      attribute float sampleVelocity;
-      attribute float previousSampleValue;
-      attribute float previousSampleVelocity;
-      uniform float mode;
-      uniform float centerX;
-      uniform float centerY;
-      uniform float scale;
-      uniform float mystery;
-      uniform float signalTime;
-      uniform float beatPulse;
-      uniform float trebleAtt;
-      uniform float previousCenterX;
-      uniform float previousCenterY;
-      uniform float previousScale;
-      uniform float previousMystery;
-      uniform float previousSignalTime;
-      uniform float previousBeatPulse;
-      uniform float previousTrebleAtt;
-      uniform float blendMix;
-      uniform float interactionOffsetX;
-      uniform float interactionOffsetY;
-      uniform float interactionRotation;
-      uniform float interactionScale;
-      ${PROCEDURAL_INTERACTION_SHADER_CHUNK}
-
-      vec2 milkdropWavePoint(
-        float t,
-        float sampleValue,
-        float velocity,
-        float pointCenterX,
-        float pointCenterY,
-        float pointScale,
-        float pointMystery,
-        float pointSignalTime,
-        float pointBeatPulse,
-        float pointTrebleAtt
-      ) {
-        float centeredSample = sampleValue - 0.5;
-        float mysteryPhase = pointMystery * 3.141592653589793;
-        float x = 0.0;
-        float y = 0.0;
-
-        if (mode < 0.5) {
-          x = -1.1 + t * 2.2;
-          y =
-            pointCenterY +
-            sin(
-              t * 3.141592653589793 * 2.0 +
-                pointSignalTime * (0.55 + pointMystery)
-            ) * (0.06 + pointTrebleAtt * 0.08) +
-            centeredSample * pointScale * 1.7 +
-            velocity * 0.12;
-        } else if (mode < 1.5) {
-          float angle =
-            t * 3.141592653589793 * 2.0 +
-            pointSignalTime * 0.32 +
-            centeredSample * 0.8 +
-            velocity * 2.5;
-          float radius =
-            0.22 +
-            sampleValue * pointScale +
-            pointBeatPulse * 0.08 +
-            sin(t * 3.141592653589793 * 4.0 + pointSignalTime) * 0.015;
-          x = pointCenterX + cos(angle) * radius;
-          y = pointCenterY + sin(angle) * radius;
-        } else if (mode < 2.5) {
-          float angle =
-            t * 3.141592653589793 * 5.0 +
-            pointSignalTime * (0.4 + pointMystery * 0.2) +
-            centeredSample * 0.65;
-          float radius =
-            0.08 + t * 0.6 + sampleValue * pointScale * 0.6 + velocity * 0.12;
-          x = pointCenterX + cos(angle) * radius;
-          y = pointCenterY + sin(angle) * radius;
-        } else if (mode < 3.5) {
-          float angle = t * 3.141592653589793 * 2.0 + pointSignalTime * 0.22;
-          float spoke =
-            0.2 +
-            sampleValue * pointScale * 1.05 +
-            sin(t * 3.141592653589793 * 12.0 + mysteryPhase) * 0.05 +
-            velocity * 0.09;
-          float pinch =
-            0.55 + cos(t * 3.141592653589793 * 6.0 + pointSignalTime) * 0.2;
-          x = pointCenterX + cos(angle) * spoke;
-          y = pointCenterY + sin(angle) * spoke * pinch;
-        } else if (mode < 4.5) {
-          x =
-            pointCenterX +
-            (sampleValue - 0.5) * pointScale * 1.85 +
-            sin(t * 3.141592653589793 * 10.0 + pointSignalTime * 0.5) * 0.04;
-          y = 1.08 - t * 2.16 + velocity * 0.22;
-        } else if (mode < 5.5) {
-          float angle = t * 3.141592653589793 * 2.0 + pointSignalTime * 0.18;
-          float xAmp = 0.26 + sampleValue * pointScale * 0.75;
-          float yAmp = 0.18 + sampleValue * pointScale;
-          x =
-            pointCenterX +
-            sin(angle * (2.0 + pointMystery * 0.6)) * xAmp +
-            cos(angle * 4.0 + mysteryPhase) * 0.04 +
-            velocity * 0.16;
-          y =
-            pointCenterY +
-            sin(angle * (3.0 + pointMystery * 0.5) + 3.141592653589793 / 2.0) *
-              yAmp;
-        } else if (mode < 6.5) {
-          float band = (sampleValue - 0.5) * pointScale * 1.4;
-          x = -1.05 + t * 2.1;
-          y =
-            pointCenterY +
-            (mod(floor(t * 512.0), 2.0) < 0.5 ? band : -band) +
-            sin(t * 3.141592653589793 * 8.0 + pointSignalTime * 0.55) * 0.03 +
-            velocity * 0.18;
-        } else {
-          float angle =
-            t * 3.141592653589793 * 2.0 +
-            pointSignalTime * (0.24 + pointMystery * 0.1);
-          float petals =
-            3.0 + floor(clamp(pointMystery * 3.0, 0.0, 3.0) + 0.5);
-          float radius =
-            0.12 +
-            (0.2 + sampleValue * pointScale * 0.9) *
-              cos(petals * angle + mysteryPhase) +
-            velocity * 0.14;
-          x = pointCenterX + cos(angle) * radius;
-          y = pointCenterY + sin(angle) * radius;
-        }
-
-        return vec2(x, y);
-      }
-
-      void main() {
-        float blendedSampleValue = mix(
-          previousSampleValue,
-          sampleValue,
-          blendMix
-        );
-        float blendedVelocity = mix(
-          previousSampleVelocity,
-          sampleVelocity,
-          blendMix
-        );
-        vec2 point = milkdropWavePoint(
-          sampleT,
-          blendedSampleValue,
-          blendedVelocity,
-          mix(previousCenterX, centerX, blendMix),
-          mix(previousCenterY, centerY, blendMix),
-          mix(previousScale, scale, blendMix),
-          mix(previousMystery, mystery, blendMix),
-          mix(previousSignalTime, signalTime, blendMix),
-          mix(previousBeatPulse, beatPulse, blendMix),
-          mix(previousTrebleAtt, trebleAtt, blendMix)
-        );
-        point = applyMilkdropInteraction(point);
-        gl_Position = projectionMatrix * modelViewMatrix * vec4(point, 0.24, 1.0);
-      }
-    `,
-    fragmentShader: `
-      uniform vec3 tint;
-      uniform float alpha;
-      uniform float interactionAlpha;
-
-      void main() {
-        gl_FragColor = vec4(tint, alpha * interactionAlpha);
-      }
-    `,
-  });
-}
-
-function createProceduralCustomWaveMaterial() {
-  return new ShaderMaterial({
-    uniforms: {
-      centerX: { value: 0 },
-      centerY: { value: 0 },
-      scaling: { value: 1 },
-      mystery: { value: 0 },
-      signalTime: { value: 0 },
-      spectrum: { value: 0 },
-      tint: { value: new Color(1, 1, 1) },
-      alpha: { value: 1 },
-      previousCenterX: { value: 0 },
-      previousCenterY: { value: 0 },
-      previousScaling: { value: 1 },
-      previousMystery: { value: 0 },
-      previousSignalTime: { value: 0 },
-      previousSpectrum: { value: 0 },
-      blendMix: { value: 1 },
-      ...createProceduralInteractionUniformState(),
-    },
-    transparent: true,
-    vertexShader: `
-      attribute float sampleT;
-      attribute float sampleValue;
-      attribute float previousSampleValue;
-      uniform float centerX;
-      uniform float centerY;
-      uniform float scaling;
-      uniform float mystery;
-      uniform float signalTime;
-      uniform float spectrum;
-      uniform float previousCenterX;
-      uniform float previousCenterY;
-      uniform float previousScaling;
-      uniform float previousMystery;
-      uniform float previousSignalTime;
-      uniform float previousSpectrum;
-      uniform float blendMix;
-      uniform float interactionOffsetX;
-      uniform float interactionOffsetY;
-      uniform float interactionRotation;
-      uniform float interactionScale;
-      ${PROCEDURAL_INTERACTION_SHADER_CHUNK}
-
-      void main() {
-        float blendedSampleValue = mix(
-          previousSampleValue,
-          sampleValue,
-          blendMix
-        );
-        float blendedCenterX = mix(previousCenterX, centerX, blendMix);
-        float blendedCenterY = mix(previousCenterY, centerY, blendMix);
-        float blendedScaling = mix(previousScaling, scaling, blendMix);
-        float blendedMystery = mix(previousMystery, mystery, blendMix);
-        float blendedSignalTime = mix(previousSignalTime, signalTime, blendMix);
-        float blendedSpectrum = mix(previousSpectrum, spectrum, blendMix);
-        float x = blendedCenterX + (-1.0 + sampleT * 2.0) * 0.85;
-        float baseY =
-          blendedCenterY +
-          (blendedSampleValue - 0.5) *
-            0.55 *
-            blendedScaling *
-            (1.0 + blendedMystery * 0.25);
-        float orbitalY =
-          blendedCenterY +
-          sin(
-            sampleT * 3.141592653589793 * 2.0 * (1.0 + blendedMystery) +
-              blendedSignalTime
-          ) *
-            0.18 *
-            blendedScaling;
-        float y = mix(orbitalY, baseY, blendedSpectrum);
-        vec2 point = applyMilkdropInteraction(vec2(x, y));
-        gl_Position = projectionMatrix * modelViewMatrix * vec4(point, 0.28, 1.0);
-      }
-    `,
-    fragmentShader: `
-      uniform vec3 tint;
-      uniform float alpha;
-      uniform float interactionAlpha;
-
-      void main() {
-        gl_FragColor = vec4(tint, alpha * interactionAlpha);
-      }
-    `,
-  });
-}
 function getShaderTextureSourceId(source: string) {
   switch (source) {
     case 'noise':
@@ -1519,116 +406,6 @@ function getUnitPolygonClosedLineGeometry(sides: number) {
   return geometry;
 }
 
-function getProceduralMeshGeometry(density: number) {
-  const safeDensity = Math.max(2, Math.round(density));
-  const cached = proceduralMeshGeometryCache.get(safeDensity);
-  if (cached) {
-    return cached;
-  }
-
-  const sourcePositions: number[] = [];
-  for (let row = 0; row < safeDensity; row += 1) {
-    for (let col = 0; col < safeDensity; col += 1) {
-      const x = (col / Math.max(1, safeDensity - 1)) * 2 - 1;
-      const y = (row / Math.max(1, safeDensity - 1)) * 2 - 1;
-
-      if (col + 1 < safeDensity) {
-        const nextX = ((col + 1) / Math.max(1, safeDensity - 1)) * 2 - 1;
-        sourcePositions.push(x, y, -0.25, nextX, y, -0.25);
-      }
-
-      if (row + 1 < safeDensity) {
-        const nextY = ((row + 1) / Math.max(1, safeDensity - 1)) * 2 - 1;
-        sourcePositions.push(x, y, -0.25, x, nextY, -0.25);
-      }
-    }
-  }
-
-  const geometry = markSharedGeometry(new BufferGeometry());
-  geometry.setAttribute(
-    'position',
-    new Float32BufferAttribute(sourcePositions, 3),
-  );
-  geometry.setAttribute(
-    'sourcePosition',
-    new Float32BufferAttribute(sourcePositions, 3),
-  );
-  setSharedGeometryBounds(geometry, { radius: PROCEDURAL_MESH_BOUNDS_RADIUS });
-  proceduralMeshGeometryCache.set(safeDensity, geometry);
-  return geometry;
-}
-
-function getProceduralMotionVectorGeometry(countX: number, countY: number) {
-  const safeCountX = Math.max(1, Math.round(countX));
-  const safeCountY = Math.max(1, Math.round(countY));
-  const cacheKey = `${safeCountX}x${safeCountY}`;
-  const cached = proceduralMotionVectorGeometryCache.get(cacheKey);
-  if (cached) {
-    return cached;
-  }
-
-  const sourcePositions: number[] = [];
-  const endpointWeights: number[] = [];
-  for (let row = 0; row < safeCountY; row += 1) {
-    for (let col = 0; col < safeCountX; col += 1) {
-      const sourceX = safeCountX === 1 ? 0 : (col / (safeCountX - 1)) * 2 - 1;
-      const sourceY = safeCountY === 1 ? 0 : (row / (safeCountY - 1)) * 2 - 1;
-      sourcePositions.push(sourceX, sourceY, 0.18, sourceX, sourceY, 0.18);
-      endpointWeights.push(0, 1);
-    }
-  }
-
-  const geometry = markSharedGeometry(new BufferGeometry());
-  geometry.setAttribute(
-    'position',
-    new Float32BufferAttribute(sourcePositions, 3),
-  );
-  geometry.setAttribute(
-    'sourcePosition',
-    new Float32BufferAttribute(sourcePositions, 3),
-  );
-  geometry.setAttribute(
-    'endpointWeight',
-    new Float32BufferAttribute(endpointWeights, 1),
-  );
-  setSharedGeometryBounds(geometry, {
-    radius: PROCEDURAL_MOTION_VECTOR_BOUNDS_RADIUS,
-  });
-  proceduralMotionVectorGeometryCache.set(cacheKey, geometry);
-  return geometry;
-}
-
-function getProceduralWaveGeometry(sampleCount: number) {
-  const safeCount = Math.max(2, Math.round(sampleCount));
-  const cached = proceduralWaveGeometryCache.get(safeCount);
-  if (cached) {
-    return cached;
-  }
-
-  const positions = new Array(safeCount * 3).fill(0);
-  const sampleT = Array.from(
-    { length: safeCount },
-    (_, index) => index / Math.max(1, safeCount - 1),
-  );
-
-  const geometry = markSharedGeometry(new BufferGeometry());
-  geometry.setAttribute('position', new Float32BufferAttribute(positions, 3));
-  geometry.setAttribute('sampleT', new Float32BufferAttribute(sampleT, 1));
-  setSharedGeometryBounds(geometry, { radius: PROCEDURAL_WAVE_BOUNDS_RADIUS });
-  proceduralWaveGeometryCache.set(safeCount, geometry);
-  return geometry;
-}
-
-function createProceduralWaveObjectGeometry(sampleCount: number) {
-  const geometry = getProceduralWaveGeometry(sampleCount).clone();
-  setGeometryBoundingSphere(
-    geometry,
-    new Vector3(0, 0, 0),
-    PROCEDURAL_WAVE_BOUNDS_RADIUS,
-  );
-  return geometry;
-}
-
 function closeLinePositions(positions: number[]) {
   if (positions.length < 6) {
     return positions;
@@ -1697,380 +474,8 @@ function getShapeFillFallbackColor(shape: MilkdropShapeVisual) {
   };
 }
 
-function syncProceduralFieldUniforms(
-  material: ShaderMaterial,
-  {
-    zoom,
-    zoomExponent,
-    rotation,
-    warp,
-    warpAnimSpeed,
-    centerX,
-    centerY,
-    scaleX,
-    scaleY,
-    translateX,
-    translateY,
-    time,
-    trebleAtt,
-    signals,
-    tint,
-    alpha,
-  }: ProceduralFieldVisualWithSignals & {
-    time: number;
-    trebleAtt: number;
-    tint: { r: number; g: number; b: number };
-    alpha: number;
-  },
-) {
-  material.uniforms.zoom.value = zoom;
-  material.uniforms.zoomExponent.value = zoomExponent;
-  material.uniforms.rotation.value = rotation;
-  material.uniforms.warp.value = warp;
-  material.uniforms.warpAnimSpeed.value = warpAnimSpeed;
-  material.uniforms.centerX.value = centerX;
-  material.uniforms.centerY.value = centerY;
-  material.uniforms.scaleX.value = scaleX;
-  material.uniforms.scaleY.value = scaleY;
-  material.uniforms.translateX.value = translateX;
-  material.uniforms.translateY.value = translateY;
-  material.uniforms.time.value = time;
-  material.uniforms.trebleAtt.value = trebleAtt;
-  material.uniforms.tint.value.setRGB(tint.r, tint.g, tint.b);
-  material.uniforms.alpha.value = alpha;
-  material.uniforms.signalTime.value = signals.time;
-  material.uniforms.signalFrame.value = signals.frame;
-  material.uniforms.signalFps.value = signals.fps;
-  material.uniforms.signalBass.value = signals.bass;
-  material.uniforms.signalMid.value = signals.mid;
-  material.uniforms.signalMids.value = signals.mids;
-  material.uniforms.signalTreble.value = signals.treble;
-  material.uniforms.signalBassAtt.value = signals.bassAtt;
-  material.uniforms.signalMidAtt.value = signals.midAtt;
-  material.uniforms.signalMidsAtt.value = signals.midsAtt;
-  material.uniforms.signalTrebleAtt.value = signals.trebleAtt;
-  material.uniforms.signalBeat.value = signals.beat;
-  material.uniforms.signalBeatPulse.value = signals.beatPulse;
-  material.uniforms.signalRms.value = signals.rms;
-  material.uniforms.signalVol.value = signals.vol;
-  material.uniforms.signalMusic.value = signals.music;
-  material.uniforms.signalWeightedEnergy.value = signals.weightedEnergy;
-}
-
-function syncProceduralInteractionUniforms(
-  material: ShaderMaterial,
-  transform: MilkdropGpuInteractionTransform | null | undefined,
-) {
-  material.uniforms.interactionOffsetX.value = transform?.offsetX ?? 0;
-  material.uniforms.interactionOffsetY.value = transform?.offsetY ?? 0;
-  material.uniforms.interactionRotation.value = transform?.rotation ?? 0;
-  material.uniforms.interactionScale.value = transform?.scale ?? 1;
-  material.uniforms.interactionAlpha.value = transform?.alphaMultiplier ?? 1;
-}
-
-function setOrUpdateScalarAttribute(
-  geometry: BufferGeometry,
-  name: string,
-  values: number[],
-) {
-  const existing = geometry.getAttribute(name);
-  if (
-    existing instanceof Float32BufferAttribute &&
-    existing.itemSize === 1 &&
-    existing.array.length === values.length
-  ) {
-    existing.array.set(values);
-    existing.needsUpdate = true;
-    return;
-  }
-  const attribute = new Float32BufferAttribute(values, 1);
-  attribute.setUsage(DynamicDrawUsage);
-  geometry.setAttribute(name, attribute);
-}
-
-function resampleScalarValues(values: number[], targetLength: number) {
-  if (values.length === targetLength) {
-    return values;
-  }
-  if (targetLength <= 0) {
-    return [];
-  }
-  if (values.length === 0) {
-    return new Array<number>(targetLength).fill(0);
-  }
-  if (values.length === 1) {
-    return new Array<number>(targetLength).fill(values[0] ?? 0);
-  }
-
-  const resampled = new Array<number>(targetLength);
-  const sourceMaxIndex = values.length - 1;
-  const targetMaxIndex = Math.max(1, targetLength - 1);
-  for (let index = 0; index < targetLength; index += 1) {
-    const position = (index / targetMaxIndex) * sourceMaxIndex;
-    const lowerIndex = Math.floor(position);
-    const upperIndex = Math.min(sourceMaxIndex, lowerIndex + 1);
-    const mix = position - lowerIndex;
-    const lowerValue = values[lowerIndex] ?? 0;
-    const upperValue = values[upperIndex] ?? lowerValue;
-    resampled[index] = lerpNumber(lowerValue, upperValue, mix);
-  }
-  return resampled;
-}
-
 function lerpNumber(previous: number, current: number, mix: number) {
   return previous + (current - previous) * mix;
-}
-
-function syncPreviousProceduralFieldUniforms(
-  material: ShaderMaterial,
-  field: ProceduralFieldVisualWithSignals,
-) {
-  material.uniforms.previousZoom.value = field.zoom;
-  material.uniforms.previousZoomExponent.value = field.zoomExponent;
-  material.uniforms.previousRotation.value = field.rotation;
-  material.uniforms.previousWarp.value = field.warp;
-  material.uniforms.previousWarpAnimSpeed.value = field.warpAnimSpeed;
-  material.uniforms.previousCenterX.value = field.centerX;
-  material.uniforms.previousCenterY.value = field.centerY;
-  material.uniforms.previousScaleX.value = field.scaleX;
-  material.uniforms.previousScaleY.value = field.scaleY;
-  material.uniforms.previousTranslateX.value = field.translateX;
-  material.uniforms.previousTranslateY.value = field.translateY;
-  material.uniforms.previousSignalTime.value = field.signals.time;
-  material.uniforms.previousSignalFrame.value = field.signals.frame;
-  material.uniforms.previousSignalFps.value = field.signals.fps;
-  material.uniforms.previousSignalBass.value = field.signals.bass;
-  material.uniforms.previousSignalMid.value = field.signals.mid;
-  material.uniforms.previousSignalMids.value = field.signals.mids;
-  material.uniforms.previousSignalTreble.value = field.signals.treble;
-  material.uniforms.previousSignalBassAtt.value = field.signals.bassAtt;
-  material.uniforms.previousSignalMidAtt.value = field.signals.midAtt;
-  material.uniforms.previousSignalMidsAtt.value = field.signals.midsAtt;
-  material.uniforms.previousSignalTrebleAtt.value = field.signals.trebleAtt;
-  material.uniforms.previousSignalBeat.value = field.signals.beat;
-  material.uniforms.previousSignalBeatPulse.value = field.signals.beatPulse;
-  material.uniforms.previousSignalRms.value = field.signals.rms;
-  material.uniforms.previousSignalVol.value = field.signals.vol;
-  material.uniforms.previousSignalMusic.value = field.signals.music;
-  material.uniforms.previousSignalWeightedEnergy.value =
-    field.signals.weightedEnergy;
-}
-
-function syncProceduralWaveObject(
-  object: Line | undefined,
-  wave: MilkdropProceduralWaveVisual,
-  interaction?: MilkdropGpuInteractionTransform | null,
-) {
-  const next =
-    object ??
-    new Line(
-      createProceduralWaveObjectGeometry(wave.samples.length),
-      createProceduralWaveMaterial(),
-    );
-  if (!(next.material instanceof ShaderMaterial)) {
-    if ('material' in next) {
-      disposeMaterial(next.material);
-    }
-    next.material = createProceduralWaveMaterial();
-  }
-
-  const sampleTAttribute = next.geometry.getAttribute('sampleT');
-  if (
-    !(
-      sampleTAttribute instanceof Float32BufferAttribute &&
-      sampleTAttribute.array.length === wave.samples.length
-    )
-  ) {
-    if (!isSharedGeometry(next.geometry)) {
-      disposeGeometry(next.geometry);
-    }
-    next.geometry = createProceduralWaveObjectGeometry(wave.samples.length);
-  }
-
-  setOrUpdateScalarAttribute(next.geometry, 'sampleValue', wave.samples);
-  setOrUpdateScalarAttribute(
-    next.geometry,
-    'previousSampleValue',
-    wave.samples,
-  );
-  setOrUpdateScalarAttribute(next.geometry, 'sampleVelocity', wave.velocities);
-  setOrUpdateScalarAttribute(
-    next.geometry,
-    'previousSampleVelocity',
-    wave.velocities,
-  );
-
-  const material = next.material as ShaderMaterial;
-  material.uniforms.mode.value = wave.mode;
-  material.uniforms.centerX.value = wave.centerX;
-  material.uniforms.centerY.value = wave.centerY;
-  material.uniforms.scale.value = wave.scale;
-  material.uniforms.mystery.value = wave.mystery;
-  material.uniforms.signalTime.value = wave.time;
-  material.uniforms.beatPulse.value = wave.beatPulse;
-  material.uniforms.trebleAtt.value = wave.trebleAtt;
-  material.uniforms.previousCenterX.value = wave.centerX;
-  material.uniforms.previousCenterY.value = wave.centerY;
-  material.uniforms.previousScale.value = wave.scale;
-  material.uniforms.previousMystery.value = wave.mystery;
-  material.uniforms.previousSignalTime.value = wave.time;
-  material.uniforms.previousBeatPulse.value = wave.beatPulse;
-  material.uniforms.previousTrebleAtt.value = wave.trebleAtt;
-  material.uniforms.blendMix.value = 1;
-  material.uniforms.tint.value.setRGB(wave.color.r, wave.color.g, wave.color.b);
-  material.uniforms.alpha.value = wave.alpha;
-  syncProceduralInteractionUniforms(material, interaction);
-  material.blending = wave.additive ? AdditiveBlending : NormalBlending;
-  return next;
-}
-
-function syncProceduralCustomWaveObject(
-  object: Line | undefined,
-  wave: MilkdropProceduralCustomWaveVisual,
-  interaction?: MilkdropGpuInteractionTransform | null,
-) {
-  const next =
-    object ??
-    new Line(
-      createProceduralWaveObjectGeometry(wave.samples.length),
-      createProceduralCustomWaveMaterial(),
-    );
-  if (!(next.material instanceof ShaderMaterial)) {
-    if ('material' in next) {
-      disposeMaterial(next.material);
-    }
-    next.material = createProceduralCustomWaveMaterial();
-  }
-
-  const sampleTAttribute = next.geometry.getAttribute('sampleT');
-  if (
-    !(
-      sampleTAttribute instanceof Float32BufferAttribute &&
-      sampleTAttribute.array.length === wave.samples.length
-    )
-  ) {
-    if (!isSharedGeometry(next.geometry)) {
-      disposeGeometry(next.geometry);
-    }
-    next.geometry = createProceduralWaveObjectGeometry(wave.samples.length);
-  }
-
-  setOrUpdateScalarAttribute(next.geometry, 'sampleValue', wave.samples);
-  setOrUpdateScalarAttribute(
-    next.geometry,
-    'previousSampleValue',
-    wave.samples,
-  );
-
-  const material = next.material as ShaderMaterial;
-  material.uniforms.centerX.value = wave.centerX;
-  material.uniforms.centerY.value = wave.centerY;
-  material.uniforms.scaling.value = wave.scaling;
-  material.uniforms.mystery.value = wave.mystery;
-  material.uniforms.signalTime.value = wave.time;
-  material.uniforms.spectrum.value = wave.spectrum ? 1 : 0;
-  material.uniforms.previousCenterX.value = wave.centerX;
-  material.uniforms.previousCenterY.value = wave.centerY;
-  material.uniforms.previousScaling.value = wave.scaling;
-  material.uniforms.previousMystery.value = wave.mystery;
-  material.uniforms.previousSignalTime.value = wave.time;
-  material.uniforms.previousSpectrum.value = wave.spectrum ? 1 : 0;
-  material.uniforms.blendMix.value = 1;
-  material.uniforms.tint.value.setRGB(wave.color.r, wave.color.g, wave.color.b);
-  material.uniforms.alpha.value = wave.alpha;
-  syncProceduralInteractionUniforms(material, interaction);
-  material.blending = wave.additive ? AdditiveBlending : NormalBlending;
-  return next;
-}
-
-function syncInterpolatedProceduralWaveObject(
-  object: Line | undefined,
-  previousWave: MilkdropProceduralWaveVisual,
-  currentWave: MilkdropProceduralWaveVisual,
-  mix: number,
-  alphaMultiplier: number,
-  interaction: MilkdropGpuInteractionTransform | null | undefined,
-) {
-  const next = syncProceduralWaveObject(object, currentWave, interaction);
-  setOrUpdateScalarAttribute(next.geometry, 'sampleValue', currentWave.samples);
-  setOrUpdateScalarAttribute(
-    next.geometry,
-    'previousSampleValue',
-    resampleScalarValues(previousWave.samples, currentWave.samples.length),
-  );
-  setOrUpdateScalarAttribute(
-    next.geometry,
-    'sampleVelocity',
-    currentWave.velocities,
-  );
-  setOrUpdateScalarAttribute(
-    next.geometry,
-    'previousSampleVelocity',
-    resampleScalarValues(
-      previousWave.velocities,
-      currentWave.velocities.length,
-    ),
-  );
-  const material = next.material as ShaderMaterial;
-  material.uniforms.previousCenterX.value = previousWave.centerX;
-  material.uniforms.previousCenterY.value = previousWave.centerY;
-  material.uniforms.previousScale.value = previousWave.scale;
-  material.uniforms.previousMystery.value = previousWave.mystery;
-  material.uniforms.previousSignalTime.value = previousWave.time;
-  material.uniforms.previousBeatPulse.value = previousWave.beatPulse;
-  material.uniforms.previousTrebleAtt.value = previousWave.trebleAtt;
-  material.uniforms.blendMix.value = mix;
-  material.uniforms.tint.value.setRGB(
-    lerpNumber(previousWave.color.r, currentWave.color.r, mix),
-    lerpNumber(previousWave.color.g, currentWave.color.g, mix),
-    lerpNumber(previousWave.color.b, currentWave.color.b, mix),
-  );
-  material.uniforms.alpha.value =
-    lerpNumber(previousWave.alpha, currentWave.alpha, mix) * alphaMultiplier;
-  material.blending =
-    previousWave.additive || currentWave.additive
-      ? AdditiveBlending
-      : NormalBlending;
-  syncProceduralInteractionUniforms(material, interaction);
-  return next;
-}
-
-function syncInterpolatedProceduralCustomWaveObject(
-  object: Line | undefined,
-  previousWave: MilkdropProceduralCustomWaveVisual,
-  currentWave: MilkdropProceduralCustomWaveVisual,
-  mix: number,
-  alphaMultiplier: number,
-  interaction: MilkdropGpuInteractionTransform | null | undefined,
-) {
-  const next = syncProceduralCustomWaveObject(object, currentWave, interaction);
-  setOrUpdateScalarAttribute(next.geometry, 'sampleValue', currentWave.samples);
-  setOrUpdateScalarAttribute(
-    next.geometry,
-    'previousSampleValue',
-    resampleScalarValues(previousWave.samples, currentWave.samples.length),
-  );
-  const material = next.material as ShaderMaterial;
-  material.uniforms.previousCenterX.value = previousWave.centerX;
-  material.uniforms.previousCenterY.value = previousWave.centerY;
-  material.uniforms.previousScaling.value = previousWave.scaling;
-  material.uniforms.previousMystery.value = previousWave.mystery;
-  material.uniforms.previousSignalTime.value = previousWave.time;
-  material.uniforms.previousSpectrum.value = previousWave.spectrum ? 1 : 0;
-  material.uniforms.blendMix.value = mix;
-  material.uniforms.tint.value.setRGB(
-    lerpNumber(previousWave.color.r, currentWave.color.r, mix),
-    lerpNumber(previousWave.color.g, currentWave.color.g, mix),
-    lerpNumber(previousWave.color.b, currentWave.color.b, mix),
-  );
-  material.uniforms.alpha.value =
-    lerpNumber(previousWave.alpha, currentWave.alpha, mix) * alphaMultiplier;
-  material.blending =
-    previousWave.additive || currentWave.additive
-      ? AdditiveBlending
-      : NormalBlending;
-  syncProceduralInteractionUniforms(material, interaction);
-  return next;
 }
 
 function lerpColor(
@@ -2194,704 +599,6 @@ function trimGroupChildren(group: Group, keepCount: number) {
     disposeObject(child as { children?: unknown[] });
     group.remove(child);
   }
-}
-
-function createWaveObject(
-  wave: MilkdropWaveVisual | null,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier = 1,
-) {
-  if (!wave || wave.positions.length === 0) {
-    return null;
-  }
-
-  if (wave.drawMode === 'dots') {
-    const object = new Points(
-      new BufferGeometry(),
-      new PointsMaterial({
-        size: wave.pointSize,
-        transparent: true,
-        opacity: wave.alpha * alphaMultiplier,
-        ...(wave.additive ? { blending: AdditiveBlending } : {}),
-      }),
-    );
-    ensureGeometryPositions(object.geometry, wave.positions);
-    setMaterialColor(object.material, wave.color, wave.alpha * alphaMultiplier);
-    object.position.z = 0.24;
-    return object;
-  }
-
-  const ObjectType =
-    wave.closed && behavior.useLineLoopPrimitives ? LineLoop : Line;
-  const object = new ObjectType(
-    new BufferGeometry(),
-    new LineBasicMaterial({
-      transparent: true,
-      opacity: wave.alpha * alphaMultiplier,
-      ...(wave.additive ? { blending: AdditiveBlending } : {}),
-    }),
-  );
-  ensureGeometryPositions(
-    object.geometry,
-    getWaveLinePositions(wave, behavior),
-  );
-  setMaterialColor(object.material, wave.color, wave.alpha * alphaMultiplier);
-  object.position.z = 0.24;
-  return object;
-}
-
-function createShapeObject(
-  shape: MilkdropShapeVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier = 1,
-) {
-  const group = new Group();
-  const fillMaterial =
-    shape.secondaryColor && behavior.supportsShapeGradient
-      ? new ShaderMaterial({
-          uniforms: {
-            primaryColor: {
-              value: new Color(shape.color.r, shape.color.g, shape.color.b),
-            },
-            secondaryColor: {
-              value: new Color(
-                shape.secondaryColor.r,
-                shape.secondaryColor.g,
-                shape.secondaryColor.b,
-              ),
-            },
-            primaryAlpha: {
-              value: (shape.color.a ?? 0.4) * alphaMultiplier,
-            },
-            secondaryAlpha: {
-              value: (shape.secondaryColor.a ?? 0) * alphaMultiplier,
-            },
-          },
-          transparent: true,
-          side: DoubleSide,
-          ...(shape.additive ? { blending: AdditiveBlending } : {}),
-          vertexShader: `
-          varying vec2 vLocal;
-          void main() {
-            vLocal = position.xy;
-            gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-          }
-        `,
-          fragmentShader: `
-          uniform vec3 primaryColor;
-          uniform vec3 secondaryColor;
-          uniform float primaryAlpha;
-          uniform float secondaryAlpha;
-          varying vec2 vLocal;
-
-          void main() {
-            float blend = clamp(length(vLocal), 0.0, 1.0);
-            vec3 color = mix(primaryColor, secondaryColor, blend);
-            float alpha = mix(primaryAlpha, secondaryAlpha, blend);
-            gl_FragColor = vec4(color, alpha);
-          }
-        `,
-        })
-      : new MeshBasicMaterial({
-          color: new Color(
-            getShapeFillFallbackColor(shape).r,
-            getShapeFillFallbackColor(shape).g,
-            getShapeFillFallbackColor(shape).b,
-          ),
-          opacity:
-            (getShapeFillFallbackColor(shape).a ?? 0.4) * alphaMultiplier,
-          transparent: true,
-          side: DoubleSide,
-          ...(shape.additive ? { blending: AdditiveBlending } : {}),
-        });
-
-  const fill = new Mesh(getUnitPolygonFillGeometry(shape.sides), fillMaterial);
-  fill.position.set(shape.x, shape.y, 0.14);
-  fill.scale.set(shape.radius, shape.radius, 1);
-  fill.rotation.z = shape.rotation;
-  group.add(fill);
-
-  if (shape.thickOutline) {
-    const accentBorder = new (behavior.useLineLoopPrimitives ? LineLoop : Line)(
-      behavior.useLineLoopPrimitives
-        ? getUnitPolygonOutlineGeometry(shape.sides)
-        : getUnitPolygonClosedLineGeometry(shape.sides),
-      new LineBasicMaterial({
-        color: new Color(
-          shape.borderColor.r,
-          shape.borderColor.g,
-          shape.borderColor.b,
-        ),
-        opacity:
-          Math.max(0.2, (shape.borderColor.a ?? 1) * 0.45) * alphaMultiplier,
-        transparent: true,
-        ...(shape.additive ? { blending: AdditiveBlending } : {}),
-      }),
-    );
-    accentBorder.position.set(shape.x, shape.y, 0.15);
-    accentBorder.scale.set(shape.radius * 1.045, shape.radius * 1.045, 1);
-    accentBorder.rotation.z = shape.rotation;
-    group.add(accentBorder);
-  }
-
-  const border = new (behavior.useLineLoopPrimitives ? LineLoop : Line)(
-    behavior.useLineLoopPrimitives
-      ? getUnitPolygonOutlineGeometry(shape.sides)
-      : getUnitPolygonClosedLineGeometry(shape.sides),
-    new LineBasicMaterial({
-      color: new Color(
-        shape.borderColor.r,
-        shape.borderColor.g,
-        shape.borderColor.b,
-      ),
-      opacity: (shape.borderColor.a ?? 1) * alphaMultiplier,
-      transparent: true,
-      ...(shape.additive ? { blending: AdditiveBlending } : {}),
-    }),
-  );
-  border.position.set(shape.x, shape.y, 0.16);
-  border.scale.set(shape.radius, shape.radius, 1);
-  border.rotation.z = shape.rotation;
-  group.add(border);
-
-  return group;
-}
-
-function syncShapeFillMaterial(
-  mesh: Mesh,
-  shape: MilkdropShapeVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier: number,
-) {
-  const wantsGradient =
-    Boolean(shape.secondaryColor) && behavior.supportsShapeGradient;
-  const existingMaterial = mesh.material;
-
-  if (wantsGradient) {
-    if (!(existingMaterial instanceof ShaderMaterial)) {
-      disposeMaterial(existingMaterial);
-      mesh.material = new ShaderMaterial({
-        uniforms: {
-          primaryColor: {
-            value: new Color(shape.color.r, shape.color.g, shape.color.b),
-          },
-          secondaryColor: {
-            value: new Color(
-              shape.secondaryColor?.r ?? 0,
-              shape.secondaryColor?.g ?? 0,
-              shape.secondaryColor?.b ?? 0,
-            ),
-          },
-          primaryAlpha: {
-            value: (shape.color.a ?? 0.4) * alphaMultiplier,
-          },
-          secondaryAlpha: {
-            value: (shape.secondaryColor?.a ?? 0) * alphaMultiplier,
-          },
-        },
-        transparent: true,
-        side: DoubleSide,
-        ...(shape.additive ? { blending: AdditiveBlending } : {}),
-        vertexShader: `
-          varying vec2 vLocal;
-          void main() {
-            vLocal = position.xy;
-            gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-          }
-        `,
-        fragmentShader: `
-          uniform vec3 primaryColor;
-          uniform vec3 secondaryColor;
-          uniform float primaryAlpha;
-          uniform float secondaryAlpha;
-          varying vec2 vLocal;
-
-          void main() {
-            float blend = clamp(length(vLocal), 0.0, 1.0);
-            vec3 color = mix(primaryColor, secondaryColor, blend);
-            float alpha = mix(primaryAlpha, secondaryAlpha, blend);
-            gl_FragColor = vec4(color, alpha);
-          }
-        `,
-      });
-    }
-
-    const material = mesh.material as ShaderMaterial;
-    material.uniforms.primaryColor.value.setRGB(
-      shape.color.r,
-      shape.color.g,
-      shape.color.b,
-    );
-    material.uniforms.secondaryColor.value.setRGB(
-      shape.secondaryColor?.r ?? 0,
-      shape.secondaryColor?.g ?? 0,
-      shape.secondaryColor?.b ?? 0,
-    );
-    material.uniforms.primaryAlpha.value =
-      (shape.color.a ?? 0.4) * alphaMultiplier;
-    material.uniforms.secondaryAlpha.value =
-      (shape.secondaryColor?.a ?? 0) * alphaMultiplier;
-    material.blending = shape.additive ? AdditiveBlending : NormalBlending;
-    return;
-  }
-
-  if (!(existingMaterial instanceof MeshBasicMaterial)) {
-    disposeMaterial(existingMaterial);
-    const fillColor = getShapeFillFallbackColor(shape);
-    mesh.material = new MeshBasicMaterial({
-      color: new Color(fillColor.r, fillColor.g, fillColor.b),
-      opacity: (fillColor.a ?? 0.4) * alphaMultiplier,
-      transparent: true,
-      side: DoubleSide,
-      ...(shape.additive ? { blending: AdditiveBlending } : {}),
-    });
-  }
-
-  const material = mesh.material as MeshBasicMaterial;
-  material.blending = shape.additive ? AdditiveBlending : NormalBlending;
-  const fillColor = getShapeFillFallbackColor(shape);
-  setMaterialColor(material, fillColor, (fillColor.a ?? 0.4) * alphaMultiplier);
-}
-
-function syncShapeOutline(
-  object: Line | LineLoop,
-  shape: MilkdropShapeVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier: number,
-  opacity: number,
-) {
-  const nextGeometry = behavior.useLineLoopPrimitives
-    ? getUnitPolygonOutlineGeometry(shape.sides)
-    : getUnitPolygonClosedLineGeometry(shape.sides);
-  if (object.geometry !== nextGeometry) {
-    object.geometry = nextGeometry;
-  }
-  object.position.set(shape.x, shape.y, 0.16);
-  object.scale.set(shape.radius, shape.radius, 1);
-  object.rotation.z = shape.rotation;
-  const material = object.material as LineBasicMaterial;
-  material.blending = shape.additive ? AdditiveBlending : NormalBlending;
-  setMaterialColor(material, shape.borderColor, opacity * alphaMultiplier);
-}
-
-function _syncShapeObject(
-  existing: Group | undefined,
-  shape: MilkdropShapeVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier: number,
-) {
-  const wantsAccent = shape.thickOutline;
-  const fillZ = 0.14;
-  const accentZ = 0.15;
-  const borderZ = 0.16;
-
-  if (!(existing instanceof Group)) {
-    if (existing) {
-      disposeObject(existing);
-    }
-    return createShapeObject(shape, behavior, alphaMultiplier);
-  }
-
-  const fill = existing.children[0];
-  const accent = existing.children[1];
-  const border = existing.children[wantsAccent ? 2 : 1];
-  const expectsLoop = behavior.useLineLoopPrimitives;
-  const hasSupportedBorder = expectsLoop
-    ? border instanceof LineLoop
-    : border instanceof Line;
-  const hasSupportedAccent = expectsLoop
-    ? accent instanceof LineLoop
-    : accent instanceof Line;
-
-  if (
-    !(fill instanceof Mesh) ||
-    !hasSupportedBorder ||
-    (wantsAccent && !hasSupportedAccent)
-  ) {
-    disposeObject(existing);
-    return createShapeObject(shape, behavior, alphaMultiplier);
-  }
-
-  if (fill.geometry !== getUnitPolygonFillGeometry(shape.sides)) {
-    fill.geometry = getUnitPolygonFillGeometry(shape.sides);
-  }
-  fill.position.set(shape.x, shape.y, fillZ);
-  fill.scale.set(shape.radius, shape.radius, 1);
-  fill.rotation.z = shape.rotation;
-  syncShapeFillMaterial(fill, shape, behavior, alphaMultiplier);
-
-  if (wantsAccent && (accent instanceof LineLoop || accent instanceof Line)) {
-    syncShapeOutline(
-      accent,
-      shape,
-      behavior,
-      alphaMultiplier,
-      Math.max(0.2, (shape.borderColor.a ?? 1) * 0.45),
-    );
-    accent.scale.set(shape.radius * 1.045, shape.radius * 1.045, 1);
-    accent.position.z = accentZ;
-  }
-
-  syncShapeOutline(
-    border as Line | LineLoop,
-    shape,
-    behavior,
-    alphaMultiplier,
-    shape.borderColor.a ?? 1,
-  );
-  border.position.z = borderZ;
-
-  if (!wantsAccent && accent) {
-    disposeObject(accent as { children?: unknown[] });
-    existing.remove(accent);
-  } else if (
-    wantsAccent &&
-    !(accent instanceof LineLoop) &&
-    !(accent instanceof Line)
-  ) {
-    const nextAccent = new (behavior.useLineLoopPrimitives ? LineLoop : Line)(
-      behavior.useLineLoopPrimitives
-        ? getUnitPolygonOutlineGeometry(shape.sides)
-        : getUnitPolygonClosedLineGeometry(shape.sides),
-      new LineBasicMaterial({
-        transparent: true,
-      }),
-    );
-    existing.add(nextAccent);
-    syncShapeOutline(
-      nextAccent,
-      shape,
-      behavior,
-      alphaMultiplier,
-      Math.max(0.2, (shape.borderColor.a ?? 1) * 0.45),
-    );
-    nextAccent.scale.set(shape.radius * 1.045, shape.radius * 1.045, 1);
-    nextAccent.position.z = accentZ;
-  }
-
-  return existing;
-}
-
-function createBorderObject(
-  border: MilkdropBorderVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier = 1,
-) {
-  const inset = border.key === 'outer' ? border.size : border.size + 0.08;
-  const left = -1 + inset * 2;
-  const right = 1 - inset * 2;
-  const top = 1 - inset * 2;
-  const bottom = -1 + inset * 2;
-  const group = markAlwaysOnscreen(new Group());
-  const fillShape = new Shape();
-  fillShape.moveTo(-1, 1);
-  fillShape.lineTo(1, 1);
-  fillShape.lineTo(1, -1);
-  fillShape.lineTo(-1, -1);
-  fillShape.lineTo(-1, 1);
-  const hole = new Path();
-  hole.moveTo(left, top);
-  hole.lineTo(left, bottom);
-  hole.lineTo(right, bottom);
-  hole.lineTo(right, top);
-  hole.lineTo(left, top);
-  fillShape.holes.push(hole);
-
-  const fill = markAlwaysOnscreen(
-    new Mesh(
-      new ShapeGeometry(fillShape),
-      new MeshBasicMaterial({
-        transparent: true,
-        opacity: Math.max(0.08, border.alpha * 0.45) * alphaMultiplier,
-        side: DoubleSide,
-      }),
-    ),
-  );
-  setMaterialColor(
-    fill.material,
-    border.color,
-    Math.max(0.08, border.alpha * 0.45) * alphaMultiplier,
-  );
-  fill.position.z = 0.285;
-  group.add(fill);
-
-  const outline = new (behavior.useLineLoopPrimitives ? LineLoop : Line)(
-    new BufferGeometry(),
-    new LineBasicMaterial({
-      transparent: true,
-      opacity: border.alpha * alphaMultiplier,
-    }),
-  );
-  outline.frustumCulled = false;
-  ensureGeometryPositions(
-    outline.geometry,
-    getBorderLinePositions(border, 0.3, behavior),
-  );
-  setMaterialColor(
-    outline.material,
-    border.color,
-    border.alpha * alphaMultiplier,
-  );
-  outline.position.z = 0.3;
-  group.add(outline);
-
-  if (!border.styled) {
-    return group;
-  }
-
-  const accent = new (behavior.useLineLoopPrimitives ? LineLoop : Line)(
-    new BufferGeometry(),
-    new LineBasicMaterial({
-      transparent: true,
-      opacity: Math.max(0.15, border.alpha * 0.55) * alphaMultiplier,
-    }),
-  );
-  accent.frustumCulled = false;
-  ensureGeometryPositions(
-    accent.geometry,
-    getBorderLinePositions(border, 0.3, behavior),
-  );
-  setMaterialColor(
-    accent.material,
-    border.color,
-    border.alpha * alphaMultiplier,
-  );
-  accent.scale.set(
-    border.key === 'outer' ? 0.985 : 1.015,
-    border.key === 'outer' ? 0.985 : 1.015,
-    1,
-  );
-  accent.position.z = 0.31;
-  group.add(accent);
-  return group;
-}
-
-function updateWaveObject(
-  object: Line | LineLoop | Points,
-  wave: MilkdropWaveVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier: number,
-) {
-  ensureGeometryPositions(
-    object.geometry,
-    getWaveLinePositions(wave, behavior),
-  );
-  if (object instanceof Points) {
-    const material = object.material as PointsMaterial;
-    material.size = wave.pointSize;
-    material.blending = wave.additive ? AdditiveBlending : NormalBlending;
-    setMaterialColor(material, wave.color, wave.alpha * alphaMultiplier);
-  } else {
-    const material = object.material as LineBasicMaterial;
-    material.blending = wave.additive ? AdditiveBlending : NormalBlending;
-    setMaterialColor(material, wave.color, wave.alpha * alphaMultiplier);
-  }
-  object.position.z = 0.24;
-}
-
-function _syncWaveObject(
-  existing: Line | LineLoop | Points | undefined,
-  wave: MilkdropWaveVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier: number,
-) {
-  const wantsPoints = wave.drawMode === 'dots';
-  const wantsLoop =
-    wave.closed && !wantsPoints && behavior.useLineLoopPrimitives;
-  const matches =
-    !!existing &&
-    ((wantsPoints && existing instanceof Points) ||
-      (wantsLoop && existing instanceof LineLoop) ||
-      (!wantsPoints && !wantsLoop && existing instanceof Line));
-
-  if (!matches) {
-    if (existing) {
-      disposeObject(existing);
-    }
-    const created = createWaveObject(wave, behavior, alphaMultiplier);
-    return created;
-  }
-
-  updateWaveObject(existing, wave, behavior, alphaMultiplier);
-  return existing;
-}
-
-function createLineObject(
-  positions: number[],
-  color: MilkdropColor,
-  alpha: number,
-  additive: boolean,
-) {
-  const object = markAlwaysOnscreen(
-    new Line(
-      new BufferGeometry(),
-      new LineBasicMaterial({
-        transparent: true,
-        opacity: alpha,
-        ...(additive ? { blending: AdditiveBlending } : {}),
-      }),
-    ),
-  );
-  ensureGeometryPositions(object.geometry, positions);
-  setMaterialColor(object.material, color, alpha);
-  object.position.z = 0.24;
-  return object;
-}
-
-function _syncLineObject(
-  existing: Line | undefined,
-  {
-    positions,
-    color,
-    alpha,
-    additive,
-  }: {
-    positions: number[];
-    color: MilkdropColor;
-    alpha: number;
-    additive: boolean;
-  },
-  alphaMultiplier: number,
-) {
-  if (!(existing instanceof Line) || existing instanceof LineLoop) {
-    if (existing) {
-      disposeObject(existing);
-    }
-    return createLineObject(
-      positions,
-      color,
-      alpha * alphaMultiplier,
-      additive,
-    );
-  }
-
-  ensureGeometryPositions(existing.geometry, positions);
-  const material = existing.material as LineBasicMaterial;
-  material.blending = additive ? AdditiveBlending : NormalBlending;
-  setMaterialColor(material, color, alpha * alphaMultiplier);
-  existing.position.z = 0.24;
-  return existing;
-}
-
-function updateBorderLine(
-  object: Line | LineLoop,
-  border: MilkdropBorderVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier: number,
-) {
-  const inset = border.key === 'outer' ? border.size : border.size + 0.08;
-  const previousInset = object.userData.borderInset as number | undefined;
-  if (previousInset !== inset) {
-    ensureGeometryPositions(
-      object.geometry,
-      getBorderLinePositions(border, 0.3, behavior),
-    );
-    object.userData.borderInset = inset;
-  }
-  setMaterialColor(
-    object.material as LineBasicMaterial,
-    border.color,
-    border.alpha * alphaMultiplier,
-  );
-}
-
-function updateBorderFill(
-  object: Mesh,
-  border: MilkdropBorderVisual,
-  alphaMultiplier: number,
-) {
-  const inset = border.key === 'outer' ? border.size : border.size + 0.08;
-  const previousInset = object.userData.borderInset as number | undefined;
-  if (previousInset !== inset) {
-    const left = -1 + inset * 2;
-    const right = 1 - inset * 2;
-    const top = 1 - inset * 2;
-    const bottom = -1 + inset * 2;
-    const fillShape = new Shape();
-    fillShape.moveTo(-1, 1);
-    fillShape.lineTo(1, 1);
-    fillShape.lineTo(1, -1);
-    fillShape.lineTo(-1, -1);
-    fillShape.lineTo(-1, 1);
-    const hole = new Path();
-    hole.moveTo(left, top);
-    hole.lineTo(left, bottom);
-    hole.lineTo(right, bottom);
-    hole.lineTo(right, top);
-    hole.lineTo(left, top);
-    fillShape.holes.push(hole);
-
-    if (!isSharedGeometry(object.geometry)) {
-      disposeGeometry(object.geometry);
-    }
-    object.geometry = new ShapeGeometry(fillShape);
-    object.userData.borderInset = inset;
-  }
-  setMaterialColor(
-    object.material as MeshBasicMaterial,
-    border.color,
-    Math.max(0.08, border.alpha * 0.45) * alphaMultiplier,
-  );
-  object.position.z = 0.285;
-}
-
-function _syncBorderObject(
-  existing: Group | undefined,
-  border: MilkdropBorderVisual,
-  behavior: MilkdropBackendBehavior,
-  alphaMultiplier: number,
-) {
-  if (!(existing instanceof Group)) {
-    if (existing) {
-      disposeObject(existing);
-    }
-    return createBorderObject(border, behavior, alphaMultiplier);
-  }
-
-  const fill = existing.children[0];
-  const outline = existing.children[1];
-  const accent = existing.children[2];
-  const wantsAccent = border.styled;
-  const expectsLoop = behavior.useLineLoopPrimitives;
-  const hasSupportedOutline = expectsLoop
-    ? outline instanceof LineLoop
-    : outline instanceof Line;
-  const hasSupportedAccent = expectsLoop
-    ? accent instanceof LineLoop
-    : accent instanceof Line;
-  if (
-    !(fill instanceof Mesh) ||
-    !hasSupportedOutline ||
-    (wantsAccent && !hasSupportedAccent)
-  ) {
-    disposeObject(existing);
-    return createBorderObject(border, behavior, alphaMultiplier);
-  }
-
-  updateBorderFill(fill, border, alphaMultiplier);
-  updateBorderLine(
-    outline as Line | LineLoop,
-    border,
-    behavior,
-    alphaMultiplier,
-  );
-  outline.position.z = 0.3;
-  if (wantsAccent && (accent instanceof LineLoop || accent instanceof Line)) {
-    updateBorderLine(accent, border, behavior, alphaMultiplier);
-    accent.scale.set(
-      border.key === 'outer' ? 0.985 : 1.015,
-      border.key === 'outer' ? 0.985 : 1.015,
-      1,
-    );
-    accent.position.z = 0.31;
-    (accent.material as LineBasicMaterial).opacity =
-      Math.max(0.15, border.alpha * 0.55) * alphaMultiplier;
-  }
-  if (!wantsAccent && accent) {
-    disposeObject(accent as { children?: unknown[] });
-    existing.remove(accent);
-  }
-  return existing;
 }
 
 class ThreeMilkdropAdapter implements MilkdropRendererAdapter {
@@ -3410,55 +1117,18 @@ class ThreeMilkdropAdapter implements MilkdropRendererAdapter {
     signals: MilkdropRenderPayload['frameState']['signals'],
     interaction?: MilkdropGpuInteractionTransform | null,
   ) {
-    const proceduralMesh =
-      this.backend === 'webgpu' &&
-      this.webgpuDescriptorPlan?.proceduralMesh !== null
-        ? gpuGeometry.meshField
-        : null;
-    if (proceduralMesh) {
-      const fieldProgramSignature =
-        proceduralMesh.program?.signature ?? 'default';
-      if (
-        !(this.meshLines.material instanceof ShaderMaterial) ||
-        this.meshLines.material.userData.fieldProgramSignature !==
-          fieldProgramSignature
-      ) {
-        disposeMaterial(this.meshLines.material);
-        this.meshLines.material = createProceduralMeshMaterial(
-          proceduralMesh.program,
-        );
-      }
-      this.meshLines.geometry = getProceduralMeshGeometry(
-        proceduralMesh.density,
-      );
-      syncProceduralFieldUniforms(this.meshLines.material as ShaderMaterial, {
-        ...proceduralMesh,
-        time: signals.time,
-        trebleAtt: signals.trebleAtt,
-        tint: mesh.color,
-        alpha: mesh.alpha,
-      });
-      syncProceduralInteractionUniforms(
-        this.meshLines.material as ShaderMaterial,
-        interaction,
-      );
-      this.meshLines.visible = mesh.alpha > 0.001;
-      return;
-    }
-
-    if (!(this.meshLines.material instanceof LineBasicMaterial)) {
-      disposeMaterial(this.meshLines.material);
-      this.meshLines.material = new LineBasicMaterial({
-        color: 0x4d66f2,
-        transparent: true,
-        opacity: 0.24,
-      });
-    }
-
-    const meshMaterial = this.meshLines.material as LineBasicMaterial;
-    ensureGeometryPositions(this.meshLines.geometry, mesh.positions);
-    setMaterialColor(meshMaterial, mesh.color, mesh.alpha);
-    this.meshLines.visible = mesh.positions.length > 0;
+    return renderMeshHelper({
+      backend: this.backend,
+      meshLines: this.meshLines,
+      mesh,
+      gpuGeometry,
+      signals,
+      webgpuDescriptorPlan: this.webgpuDescriptorPlan,
+      interaction,
+      disposeMaterial,
+      ensureGeometryPositions,
+      setMaterialColor,
+    });
   }
 
   private renderMotionVectors(
@@ -3483,12 +1153,6 @@ class ThreeMilkdropAdapter implements MilkdropRendererAdapter {
       cpuGroup,
       proceduralObject,
       clearGroup,
-      disposeMaterial,
-      createProceduralMotionVectorMaterial,
-      getProceduralMotionVectorGeometry,
-      syncProceduralFieldUniforms,
-      syncProceduralInteractionUniforms,
-      syncPreviousProceduralFieldUniforms,
       renderLineVisualGroup: (target, group, lines, nextAlphaMultiplier) =>
         this.renderLineVisualGroup(target, group, lines, nextAlphaMultiplier),
     });
