+++ assets/js/milkdrop/renderer-helpers/motion-vector-renderer.ts
@@ -1,20 +1,95 @@
-import type { BufferGeometry, Group, LineSegments, Material } from 'three';
-import { type LineBasicMaterial, ShaderMaterial } from 'three';
-import type {
-  MilkdropGpuFieldProgramDescriptor,
-  MilkdropGpuInteractionTransform,
-  MilkdropProceduralFieldTransformVisual,
-  MilkdropProceduralMotionVectorFieldVisual,
-  MilkdropRenderPayload,
-} from '../types';
+import type { Group, LineSegments } from 'three';
+import {
+  BufferGeometry,
+  Float32BufferAttribute,
+  type LineBasicMaterial,
+  ShaderMaterial,
+  Sphere,
+  Vector3,
+} from 'three';
+import { createProceduralMotionVectorMaterial } from '../renderer-backends/webgpu-procedural-materials';
+import type { MilkdropRenderPayload } from '../types';
+import type { ProceduralFieldVisualWithSignals } from './procedural-field-uniforms';
+import {
+  syncPreviousProceduralFieldUniforms,
+  syncProceduralFieldUniforms,
+  syncProceduralInteractionUniforms,
+} from './procedural-field-uniforms';
 
-type ProceduralFieldUniformInput = MilkdropProceduralFieldTransformVisual & {
-  signals: MilkdropProceduralMotionVectorFieldVisual['signals'];
+const SHARED_GEOMETRY_FLAG = 'milkdropSharedGeometry';
+const PROCEDURAL_MOTION_VECTOR_BOUNDS_RADIUS = Math.SQRT2 * 2.35;
+const proceduralMotionVectorGeometryCache = new Map<string, BufferGeometry>();
+
+type ProceduralFieldUniformInput = ProceduralFieldVisualWithSignals & {
   time: number;
   trebleAtt: number;
   tint: { r: number; g: number; b: number };
   alpha: number;
 };
+
+function markSharedGeometry<T extends BufferGeometry>(geometry: T) {
+  geometry.userData[SHARED_GEOMETRY_FLAG] = true;
+  return geometry;
+}
+
+function setGeometryBoundingSphere(
+  geometry: BufferGeometry,
+  center: Vector3,
+  radius: number,
+) {
+  if (!geometry.boundingSphere) {
+    geometry.boundingSphere = new Sphere(center.clone(), radius);
+    return geometry.boundingSphere;
+  }
+  geometry.boundingSphere.center.copy(center);
+  geometry.boundingSphere.radius = radius;
+  return geometry.boundingSphere;
+}
+
+export function getProceduralMotionVectorGeometry(
+  countX: number,
+  countY: number,
+) {
+  const safeCountX = Math.max(1, Math.round(countX));
+  const safeCountY = Math.max(1, Math.round(countY));
+  const cacheKey = `${safeCountX}x${safeCountY}`;
+  const cached = proceduralMotionVectorGeometryCache.get(cacheKey);
+  if (cached) {
+    return cached;
+  }
+
+  const sourcePositions: number[] = [];
+  const endpointWeights: number[] = [];
+  for (let row = 0; row < safeCountY; row += 1) {
+    for (let col = 0; col < safeCountX; col += 1) {
+      const sourceX = safeCountX === 1 ? 0 : (col / (safeCountX - 1)) * 2 - 1;
+      const sourceY = safeCountY === 1 ? 0 : (row / (safeCountY - 1)) * 2 - 1;
+      sourcePositions.push(sourceX, sourceY, 0.18, sourceX, sourceY, 0.18);
+      endpointWeights.push(0, 1);
+    }
+  }
+
+  const geometry = markSharedGeometry(new BufferGeometry());
+  geometry.setAttribute(
+    'position',
+    new Float32BufferAttribute(sourcePositions, 3),
+  );
+  geometry.setAttribute(
+    'sourcePosition',
+    new Float32BufferAttribute(sourcePositions, 3),
+  );
+  geometry.setAttribute(
+    'endpointWeight',
+    new Float32BufferAttribute(endpointWeights, 1),
+  );
+  setGeometryBoundingSphere(
+    geometry,
+    new Vector3(0, 0, 0),
+    PROCEDURAL_MOTION_VECTOR_BOUNDS_RADIUS,
+  );
+  proceduralMotionVectorGeometryCache.set(cacheKey, geometry);
+  return geometry;
+}
 
 export function renderMotionVectors({
   backend,
@@ -26,12 +101,6 @@ export function renderMotionVectors({
   cpuGroup,
   proceduralObject,
   clearGroup,
-  disposeMaterial,
-  createProceduralMotionVectorMaterial,
-  getProceduralMotionVectorGeometry,
-  syncProceduralFieldUniforms,
-  syncProceduralInteractionUniforms,
-  syncPreviousProceduralFieldUniforms,
   renderLineVisualGroup,
 }: {
   backend: 'webgl' | 'webgpu';
@@ -46,26 +115,6 @@ export function renderMotionVectors({
     LineBasicMaterial | ShaderMaterial
   >;
   clearGroup: (group: Group) => void;
-  disposeMaterial: (material: Material | Material[] | null | undefined) => void;
-  createProceduralMotionVectorMaterial: (
-    program?: MilkdropGpuFieldProgramDescriptor | null,
-  ) => ShaderMaterial;
-  getProceduralMotionVectorGeometry: (
-    countX: number,
-    countY: number,
-  ) => BufferGeometry;
-  syncProceduralFieldUniforms: (
-    material: ShaderMaterial,
-    field: ProceduralFieldUniformInput,
-  ) => void;
-  syncProceduralInteractionUniforms: (
-    material: ShaderMaterial,
-    interaction: MilkdropGpuInteractionTransform | null | undefined,
-  ) => void;
-  syncPreviousProceduralFieldUniforms: (
-    material: ShaderMaterial,
-    field: MilkdropProceduralMotionVectorFieldVisual,
-  ) => void;
   renderLineVisualGroup: (
     target: 'trails' | 'motion-vectors' | 'blend-motion-vectors',
     group: Group,
@@ -87,7 +136,7 @@ export function renderMotionVectors({
       proceduralObject.material.userData.fieldProgramSignature !==
         fieldProgramSignature
     ) {
-      disposeMaterial(proceduralObject.material);
+      proceduralObject.material.dispose();
       proceduralObject.material = createProceduralMotionVectorMaterial(
         proceduralField.program,
       );
@@ -113,7 +162,7 @@ export function renderMotionVectors({
           ),
           1,
         ) * alphaMultiplier,
-    });
+    } satisfies ProceduralFieldUniformInput);
     const proceduralMaterial = proceduralObject.material as ShaderMaterial;
     syncProceduralInteractionUniforms(
       proceduralMaterial,
